+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog KasperskyOS
 
+## Release for KasperskyOS Community Edition SDK 1.3 - 2025-07
+
+### Changes
+
+* Updated STANDARD_GNU compiler flag to version 17
+* Added support for the Radxa ROCK 3A hardware platform
+
 ## Release for KasperskyOS Community Edition SDK 1.2 - 2024-12
 
 ### Changes